requirements.txt
numpy
pandas
matplotlib
seaborn
tensorflow
langchain
langgraph
pandas
matplotlib
langchain_core
langchain_community
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
langgraph
poetry
arxiv
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
langchainhub
sentence-transformers
langchain-objectbox
pinecone-client
pinecone-text
pinecone-notebooks